% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMISCdata.R
\docType{data}
\name{logMISCdata}
\alias{logMISCdata}
\title{Protein-level validation data set}
\format{
A test data set

\describe{
\item{\code{M.x}}{a numeric vector}
\item{\code{K.x}}{a numeric vector}
\item{\code{P.x}}{a numeric vector}
\item{\code{A.x}}{a numeric vector}
}
}
\usage{
data(logMISCdata)
}
\value{
A test data set
}
\description{
Validation data set profiles. Ad data set of protein levels for
samples of MIS-C (M.x), Kawasaki disease (K.x),
pneumonia (P.x), and asymptomatic (A.x)
}
\examples{
data(logMISCdata)
str(logMISCdata)
}
\references{
Rivera JG, Zheng H, Richlin B, Suarez C,
Gaur S, Ricciardi E, Hasan UN, Cuddy W, Singh AR,
Bukulmez H, Kaelber DC, Kimuara Y, Brady PW,
Wahezi D, Rothschild E, Lakhani SA, Herbst KW,
Hogan AH, Salazar JC, Moroso-Fela S, Roy J,
Kleinman LC, Horton DB,
Moore DF, Gennaro M(2025)
Combining Mass Spectrometry with Machine Learning to
Identify Novel Protein Signatures: The Example of Multisystem
Inflammatory Syndrome in Children
}
\keyword{datasets}
