% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMISCdataTrain.R
\docType{data}
\name{logMISCdataTrain}
\alias{logMISCdataTrain}
\title{Protein-level training data set}
\format{
The training data set

\describe{
\item{\code{A}}{a numeric vector}
\item{\code{M}}{a numeric vector}
}
}
\usage{
data(logMISCdataTrain)
}
\value{
A test data set
}
\description{
Training data set profiles. Ad data set of protein levels for
samples of asymptomatic (A) and MIS-C (M)
}
\examples{
data(logMISCdataTrain)
str(logMISCdataTrain)
}
\references{
Rivera JG, Moore DF, Gennaro M(2025)
}
\keyword{datasets}
